+++ lib/src/backend/constants.dart
@@ -1,0 +1,11 @@
+// @license
+// Copyright (c) 2019 - 2025 Dr. Gabriel Gatzsche. All Rights Reserved.
+//
+// Use of this source code is governed by terms that can be
+// found in the LICENSE file in the root of this package.
+
+/// The master workspace folder
+const String kidneyMasterFolder = '.master';
+
+/// The folder where tickets are stored
+const String kidneyTicketFolder = 'tickets';