+++ test/commands/add_deps_test.dart
@@ -9,6 +9,7 @@ import 'dart:io';
 
 import 'package:args/command_runner.dart';
 import 'package:gg_capture_print/gg_capture_print.dart';
+import 'package:kidney_core/src/backend/constants.dart';
 import 'package:kidney_core/src/backend/git_handler.dart';
 import 'package:kidney_core/src/commands/add_deps.dart';
 import 'package:mocktail/mocktail.dart';
@@ -40,7 +41,7 @@ void main() {
       when(() => mockGitCloner.cloneRepo(any(), any()))
           .thenAnswer((_) async {});
       tempDir = Directory.systemTemp.createTempSync('add_deps_test');
-      workspacePath = path.join(tempDir.path, 'kidney_ws_master');
+      workspacePath = path.join(tempDir.path, kidneyMasterFolder);
       Directory(workspacePath).createSync(recursive: true);
       dirNoPubspec = Directory(path.join(workspacePath, 'no_pubspec'))
         ..createSync(recursive: true);