+++ test/commands/list/organizations_test.dart
@@ -8,6 +8,7 @@ import 'dart:io';
 
 import 'package:args/command_runner.dart';
 import 'package:gg_capture_print/gg_capture_print.dart';
+import 'package:kidney_core/src/backend/constants.dart';
 import 'package:test/test.dart';
 import 'package:kidney_core/src/commands/list/organizations.dart';
 import 'package:path/path.dart' as path;
@@ -27,7 +28,7 @@ void main() {
     setUp(() {
       messages.clear();
       tempDir = Directory.systemTemp.createTempSync('list_org_test');
-      masterDir = Directory(path.join(tempDir.path, 'kidney_ws_master'))
+      masterDir = Directory(path.join(tempDir.path, kidneyMasterFolder))
         ..createSync(recursive: true);
     });