+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Add force flag correctly to AddCommand
 - Add command kidney\_core list tickets
 - add support in add command for multiple repos
+- Add constants.dart and change master folder to .master
 
 ### Changed